Ajoute test fonctionnel /
