+++ server/api.yaml
@@ -5,6 +5,17 @@ info:
   version: 1.0.0
 
 paths:
+  /:
+    get:
+      summary: Renvoie un message de bienvenue
+      operationId: server.handlers.Welcome.home
+      responses:
+          '200':
+            description: Un petit bonjour
+            content:
+              application/json:
+                schema:
+                  $ref: '#/components/schemas/Greetings'
   /calculate/cas_types:
     post:
       summary: Calcule l'impact par cas types
@@ -29,17 +40,6 @@ paths:
               application/json:
                 schema:
                   $ref: '#/components/schemas/revenus'
-  /:
-    get:
-      summary: Renvoie un message de bienvenue
-      operationId: server.handlers.CasTypes.hello
-      responses:
-          '200':
-            description: Un petit bonjour
-            content:
-              application/json:
-                schema:
-                  $ref: '#/components/schemas/greetings'
   /metadata/description_cas_types:
     post:
       summary: Decrit les cas types
@@ -79,6 +79,11 @@ paths:
             description: Qu'est ce que c'est description
 components:
   schemas:
+    Greetings:
+      type: object
+      properties:
+        hello:
+          type: string
     CalculateCasTypes:
       type: object
       properties:
@@ -86,11 +91,6 @@ components:
           type: object
         cas_types:
           type: array
-    greetings:
-      type: object
-      properties:
-        hello:
-          type: string
     revenus:
       type: object
     compareoldnew_old:
