+++ server/handlers/__init__.py
@@ -1,1 +1,2 @@
 from .cas_types import CasTypes, SimulationRunner  # noqa
+from .welcome import Welcome  # noqa
