+++ server/handlers/welcome.py
@@ -1,0 +1,6 @@
+# -*- coding: utf-8 -*-
+
+
+class Welcome(object):
+    def home(**params: dict) -> tuple:
+        return {"hello": "coucou"}, 200
