+++ tests/server/handlers/test_cas_types.py
@@ -3,25 +3,6 @@
 import pytest
 import json
 
-from server.app import app
-
-
-@pytest.fixture
-def client():
-    app.config['TESTING'] = True
-    client = app.test_client()
-    return client
-
-
-@pytest.fixture
-def mimetype() -> str:
-    return "application/json"
-
-
-@pytest.fixture
-def headers(mimetype: str) -> dict:
-    return {"Content-Type": mimetype, "Accept": mimetype}
-
 
 @pytest.fixture
 def data() -> str:
